pyfiles/twse/requirements.txt
absl-py==2.2.2 ; python_version == "3.11"
astunparse==1.6.3 ; python_version == "3.11"
beautifulsoup4==4.13.3 ; python_version == "3.11"
bs4==0.0.2 ; python_version == "3.11"
cachetools==5.5.2 ; python_version == "3.11"
certifi==2025.1.31 ; python_version == "3.11"
charset-normalizer==3.4.1 ; python_version == "3.11"
flatbuffers==25.2.10 ; python_version == "3.11"
gast==0.6.0 ; python_version == "3.11"
google-api-core==2.24.2 ; python_version == "3.11"
google-auth-oauthlib==1.2.1 ; python_version == "3.11"
google-auth==2.38.0 ; python_version == "3.11"
google-cloud-core==2.4.3 ; python_version == "3.11"
google-cloud-storage==3.1.0 ; python_version == "3.11"
google-crc32c==1.7.1 ; python_version == "3.11"
google-pasta==0.2.0 ; python_version == "3.11"
google-resumable-media==2.7.2 ; python_version == "3.11"
googleapis-common-protos==1.70.0 ; python_version == "3.11"
grpcio==1.71.0 ; python_version == "3.11"
h5py==3.13.0 ; python_version == "3.11"
idna==3.10 ; python_version == "3.11"
keras==2.15.0 ; python_version == "3.11"
libclang==18.1.1 ; python_version == "3.11"
markdown==3.7 ; python_version == "3.11"
markupsafe==3.0.2 ; python_version == "3.11"
ml-dtypes==0.2.0 ; python_version == "3.11"
numpy==1.26.4 ; python_version == "3.11"
oauthlib==3.2.2 ; python_version == "3.11"
opencv-python==4.11.0.86 ; python_version == "3.11"
opt-einsum==3.4.0 ; python_version == "3.11"
packaging==24.2 ; python_version == "3.11"
pandas==2.2.3 ; python_version == "3.11"
proto-plus==1.26.1 ; python_version == "3.11"
protobuf==4.25.6 ; python_version == "3.11"
pyasn1-modules==0.4.2 ; python_version == "3.11"
pyasn1==0.6.1 ; python_version == "3.11"
python-dateutil==2.9.0.post0 ; python_version == "3.11"
pytz==2025.2 ; python_version == "3.11"
requests-oauthlib==2.0.0 ; python_version == "3.11"
requests==2.32.3 ; python_version == "3.11"
rsa==4.9 ; python_version == "3.11"
setuptools==78.1.0 ; python_version == "3.11"
six==1.17.0 ; python_version == "3.11"
soupsieve==2.6 ; python_version == "3.11"
tensorboard-data-server==0.7.2 ; python_version == "3.11"
tensorboard==2.15.2 ; python_version == "3.11"
tensorflow-estimator==2.15.0 ; python_version == "3.11"
tensorflow-io-gcs-filesystem==0.37.1 ; python_version == "3.11"
tensorflow==2.15.0 ; python_version == "3.11"
termcolor==3.0.1 ; python_version == "3.11"
typing-extensions==4.13.1 ; python_version == "3.11"
tzdata==2025.2 ; python_version == "3.11"
urllib3==2.3.0 ; python_version == "3.11"
werkzeug==3.1.3 ; python_version == "3.11"
wheel==0.45.1 ; python_version == "3.11"
wrapt==1.14.1 ; python_version == "3.11"
google-cloud-firestore==2.20.2 ; python_version == "3.11"